chore(friendshipper): move OFPA cache into EngineProvider
