+++ friendshipper/src-tauri/src/engine/mod.rs
@@ -1,5 +1,6 @@
 mod provider;
 mod unreal;
 
+pub use provider::CommunicationType;
 pub use provider::EngineProvider;
 pub use unreal::UnrealEngineProvider;
