+++ friendshipper/src-tauri/src/repo/operations/log.rs
@@ -49,7 +49,7 @@ where
                 repo_status: state.repo_status.clone(),
                 app_config: state.app_config.clone(),
                 repo_config: state.repo_config.clone(),
-                ofpa_cache: state.ofpa_cache.clone(),
+                engine: state.engine.clone(),
                 git_client: state.git(),
                 aws_client: aws_client.clone(),
                 storage: state.storage.read().clone().unwrap(),
